docs: bump version and improve readme
