+++ packages/cli/README.md
@@ -84,7 +84,7 @@ Initialize a new Bucket configuration in your project.
 This creates a `bucket.config.json` file with your settings and prompts for any required information not provided via options.
 
 ```bash
-bucket init [--overwrite]
+npx bucket init [--overwrite]
 ```
 
 Options:
@@ -99,7 +99,7 @@ All-in-one command to get started quickly. This command combines `init`, feature
 and type generation in a single step. Use this for the fastest way to get up and running with Bucket.
 
 ```bash
-bucket new "My Feature" [--key my-feature] [--app-id ap123456789] [--key-format custom] [--out gen/features.ts] [--format react]
+npx bucket new "My Feature" [--key my-feature] [--app-id ap123456789] [--key-format custom] [--out gen/features.ts] [--format react]
 ```
 
 Options:
@@ -117,7 +117,7 @@ If you prefer more control over each step, you can use the individual commands (
 Log in to your Bucket account. This will authenticate your CLI for subsequent operations and store credentials securely.
 
 ```bash
-bucket login
+npx bucket login
 ```
 
 ### `bucket logout`
@@ -125,7 +125,7 @@ bucket login
 Log out from your Bucket account, removing stored credentials.
 
 ```bash
-bucket logout
+npx bucket logout
 ```
 
 ### `bucket features`
@@ -138,7 +138,7 @@ Create a new feature in your Bucket app.
 The command guides you through the feature creation process with interactive prompts if options are not provided.
 
 ```bash
-bucket features create "My Feature" [--key my-feature] [--app-id ap123456789] [--key-format custom]
+npx bucket features create "My Feature" [--key my-feature] [--app-id ap123456789] [--key-format custom]
 ```
 
 Options:
@@ -153,7 +153,7 @@ List all features for the current app.
 This helps you visualize what features are available and their current configurations.
 
 ```bash
-bucket features list [--app-id ap123456789]
+npx bucket features list [--app-id ap123456789]
 ```
 
 Options:
@@ -166,7 +166,7 @@ Generate TypeScript types for your features.
 This ensures type safety when using Bucket features in your TypeScript/JavaScript applications.
 
 ```bash
-bucket features types [--app-id ap123456789] [--out gen/features.ts] [--format react]
+npx bucket features types [--app-id ap123456789] [--out gen/features.ts] [--format react]
 ```
 
 Options:
@@ -185,7 +185,7 @@ List all companies for the current app.
 This helps you visualize the companies using your features and their basic metrics.
 
 ```bash
-bucket companies list [--app-id ap123456789] [--filter nameOrId]
+npx bucket companies list [--app-id ap123456789] [--filter nameOrId]
 ```
 
 Options:
@@ -195,27 +195,97 @@ Options:
 
 #### `bucket companies features access`
 
-Grant or revoke access to specific features for a company.
+Grant or revoke access to specific features for companies, segments, and users.
 If no feature key is provided, you'll be prompted to select one from a list.
 
 ```bash
-bucket companies features access <companyId> [featureKey] [--enable|--disable] [--app-id ap123456789]
+npx bucket companies features access [featureKey] [--enable|--disable] [--companies <id...>] [--segments <id...>] [--users <id...>] [--app-id ap123456789]
 ```
 
 Arguments:
 
-- `companyId`: ID of the company to manage
 - `featureKey`: Key of the feature to grant/revoke access to (optional, interactive selection if omitted)
 
 Options:
 
-- `--enable`: Enable the feature for this company
-- `--disable`: Disable the feature for this company
+- `--enable`: Enable the feature for the specified targets
+- `--disable`: Disable the feature for the specified targets
+- `--users`: User IDs to target. Can be specified multiple times
+- `--companies`: Company IDs to target. Can be specified multiple times
+- `--segments`: Segment IDs to target. Can be specified multiple times
 - `--app-id`: App ID to use
+
+At least one target (companies, segments, or users) must be specified. You must also specify either `--enable` or `--disable`, but not both.
+
+Example:
+
+```bash
+# Enable feature for multiple companies and users
+npx bucket companies features access my-feature --enable --companies comp_123 --companies comp_456 --users user_789
+```
 
 ### `bucket apps`
 
 Commands for managing Bucket apps.
+
+## Model Context Protocol (Beta)
+
+The Model Context Protocol (MCP) is an open protocol that provides a standardized way to connect AI models to different data sources and tools. In the context of Bucket, MCP enables your development environment to understand your feature flags, their states, and their relationships within your codebase. This creates a seamless bridge between your feature management workflow and AI-powered development tools. MCP is in a very early stage of development and changes are frequent, if something isn't working please check out the [Model Context Protocol Website](https://modelcontextprotocol.io/) and open an [issue ticket here](https://github.com/bucketco/bucket-javascript-sdk/issues).
+
+### Setting up MCP
+
+MCP servers currently run locally on your machine. To start the MCP server run the CLI command from your Bucket initialized project directory:
+
+```bash
+npx bucket mcp
+```
+
+This will start an SSE server at `http://localhost:8050/sse` by default which you can connect to using your [client of choice](https://modelcontextprotocol.io/clients). Below are examples that work for [Cursor](https://www.cursor.com/) and [Claude Desktop](https://claude.ai/download).
+
+#### Server-Side Events (SSE)
+
+```json
+{
+  "mcpServers": {
+    "Bucket": {
+      "url": "http://localhost:8050/sse"
+    }
+  }
+}
+```
+
+#### STDIO Proxy
+
+Some clients don't support SSE and can instead interface with the MCP server over a STDIO proxy.
+
+```json
+{
+  "mcpServers": {
+    "Bucket": {
+      "command": "npx",
+      "args": ["-y", "supergateway", "--sse", "http://localhost:8050/sse"]
+    }
+  }
+}
+```
+
+### Cursor IDE
+
+To enable MCP features in Cursor IDE:
+
+1. Open Cursor IDE.
+2. Go to `Settings > MCP`.
+3. Click `Add new global MCP server` and paste the `SSE` config.
+4. Save and go back to Cursor.
+
+### Clause Desktop
+
+To enable MCP features in Cursor Desktop:
+
+1. Open Claude Desktop.
+2. Go to `Settings > Developer`.
+3. Click `Edit config` and paste the `STDIO` config.
+4. Save and restart Claude Desktop.
 
 ## Global Options
