+++ packages/cli/commands/companies.ts
@@ -1,26 +1,16 @@
-import { select } from "@inquirer/prompts";
 import chalk from "chalk";
-import { Argument, Command } from "commander";
+import { Command } from "commander";
 import ora, { Ora } from "ora";
 
 import { getApp } from "../services/bootstrap.js";
 import { listCompanies } from "../services/companies.js";
-import { listFeatures, updateFeatureAccess } from "../services/features.js";
 import { configStore } from "../stores/config.js";
 import {
   handleError,
   MissingAppIdError,
   MissingEnvIdError,
 } from "../utils/errors.js";
-import {
-  appIdOption,
-  companyFilterOption,
-  companyIdsOption,
-  disableFeatureOption,
-  enableFeatureOption,
-  segmentIdsOption,
-  userIdsOption,
-} from "../utils/options.js";
+import { appIdOption, companyFilterOption } from "../utils/options.js";
 import { baseUrlSuffix } from "../utils/path.js";
 
 export const listCompaniesAction = async (options: { filter?: string }) => {
@@ -67,128 +57,9 @@ export const listCompaniesAction = async (options: { filter?: string }) => {
   }
 };
 
-export const featureAccessAction = async (
-  featureKey: string | undefined,
-  options: {
-    enable: boolean;
-    disable: boolean;
-    companies?: string[];
-    segments?: string[];
-    users?: string[];
-  },
-) => {
-  const { baseUrl, appId } = configStore.getConfig();
-  let spinner: Ora | undefined;
-
-  if (!appId) {
-    return handleError(new MissingAppIdError(), "Feature Access");
-  }
-
-  const app = getApp(appId);
-  const production = app.environments.find((e) => e.isProduction);
-  if (!production) {
-    return handleError(new MissingEnvIdError(), "Feature Access");
-  }
-
-  // Validate conflicting options
-  if (options.enable && options.disable) {
-    return handleError(
-      "Cannot both enable and disable a feature.",
-      "Feature Access",
-    );
-  }
-
-  if (!options.enable && !options.disable) {
-    return handleError(
-      "Must specify either --enable or --disable.",
-      "Feature Access",
-    );
-  }
-
-  // Validate at least one target is specified
-  if (
-    !options.companies?.length &&
-    !options.segments?.length &&
-    !options.users?.length
-  ) {
-    return handleError(
-      "Must specify at least one target using --companies, --segments, or --users.",
-      "Feature Access",
-    );
-  }
-
-  // If feature key is not provided, let user select one
-  if (!featureKey) {
-    try {
-      spinner = ora(
-        `Loading features for app ${chalk.cyan(app.name)}${baseUrlSuffix(
-          baseUrl,
-        )}...`,
-      ).start();
-
-      const featuresResponse = await listFeatures(appId, {
-        envId: production.id,
-      });
-
-      if (featuresResponse.data.length === 0) {
-        return handleError("No features found for this app.", "Feature Access");
-      }
-
-      spinner.succeed(
-        `Loaded features for app ${chalk.cyan(app.name)}${baseUrlSuffix(baseUrl)}.`,
-      );
-
-      featureKey = await select({
-        message: "Select a feature to manage access:",
-        choices: featuresResponse.data.map((feature) => ({
-          name: `${feature.name} (${feature.key})`,
-          value: feature.key,
-        })),
-      });
-    } catch (error) {
-      spinner?.fail("Loading features failed.");
-      return handleError(error, "Feature Access");
-    }
-  }
-
-  // Determine if enabling or disabling
-  const isEnabled = options.enable;
-
-  try {
-    const targetTypes = [];
-    if (options.companies?.length) targetTypes.push("companies");
-    if (options.segments?.length) targetTypes.push("segments");
-    if (options.users?.length) targetTypes.push("users");
-
-    spinner = ora(
-      `${isEnabled ? "Enabling" : "Disabling"} feature ${chalk.cyan(featureKey)} for ${targetTypes.join(", ")}...`,
-    ).start();
-
-    await updateFeatureAccess(appId, {
-      envId: production.id,
-      featureKey,
-      isEnabled,
-      companyIds: options.companies,
-      segmentIds: options.segments,
-      userIds: options.users,
-    });
-
-    spinner.succeed(
-      `${isEnabled ? "Enabled" : "Disabled"} feature ${chalk.cyan(featureKey)} for ${targetTypes.join(", ")}.`,
-    );
-  } catch (error) {
-    spinner?.fail(`Feature access update failed.`);
-    void handleError(error, "Feature Access");
-  }
-};
-
 export function registerCompanyCommands(cli: Command) {
   const companiesCommand = new Command("companies").description(
     "Manage companies.",
-  );
-
-  const featuresCommand = new Command("features").description(
-    "Manage feature access.",
   );
 
   companiesCommand
@@ -197,28 +68,6 @@ export function registerCompanyCommands(cli: Command) {
     .addOption(appIdOption)
     .addOption(companyFilterOption)
     .action(listCompaniesAction);
-
-  // Feature access command
-  featuresCommand
-    .command("access")
-    .description(
-      "Grant or revoke feature access for companies, segments, and users.",
-    )
-    .addOption(appIdOption)
-    .addArgument(
-      new Argument(
-        "[featureKey]",
-        "Feature key. If not provided, you'll be prompted to select one",
-      ),
-    )
-    .addOption(enableFeatureOption)
-    .addOption(disableFeatureOption)
-    .addOption(companyIdsOption)
-    .addOption(segmentIdsOption)
-    .addOption(userIdsOption)
-    .action(featureAccessAction);
-
-  companiesCommand.addCommand(featuresCommand);
 
   // Update the config with the cli override values
   companiesCommand.hook("preAction", (_, command) => {
