+++ packages/cli/commands/features.ts
@@ -1,11 +1,16 @@
-import { input } from "@inquirer/prompts";
+import { input, select } from "@inquirer/prompts";
 import chalk from "chalk";
-import { Command } from "commander";
+import { Argument, Command } from "commander";
 import { relative } from "node:path";
 import ora, { Ora } from "ora";
 
 import { getApp, getOrg } from "../services/bootstrap.js";
-import { createFeature, Feature, listFeatures } from "../services/features.js";
+import {
+  createFeature,
+  Feature,
+  listFeatures,
+  updateFeatureAccess,
+} from "../services/features.js";
 import { configStore } from "../stores/config.js";
 import {
   handleError,
@@ -21,12 +26,19 @@ import {
 } from "../utils/gen.js";
 import {
   appIdOption,
+  companyIdsOption,
+  disableFeatureOption,
+  enableFeatureOption,
   featureKeyOption,
   featureNameArgument,
+  segmentIdsOption,
   typesFormatOption,
   typesOutOption,
+  userIdsOption,
 } from "../utils/options.js";
 import { baseUrlSuffix, featureUrl } from "../utils/path.js";
+
+const lf = new Intl.ListFormat("en");
 
 type CreateFeatureArgs = {
   key?: string;
@@ -173,6 +185,122 @@ export const generateTypesAction = async () => {
   }
 };
 
+export const featureAccessAction = async (
+  featureKey: string | undefined,
+  options: {
+    enable: boolean;
+    disable: boolean;
+    companies?: string[];
+    segments?: string[];
+    users?: string[];
+  },
+) => {
+  const { baseUrl, appId } = configStore.getConfig();
+  let spinner: Ora | undefined;
+
+  if (!appId) {
+    return handleError(new MissingAppIdError(), "Feature Access");
+  }
+
+  const app = getApp(appId);
+  const production = app.environments.find((e) => e.isProduction);
+  if (!production) {
+    return handleError(new MissingEnvIdError(), "Feature Access");
+  }
+
+  // Validate conflicting options
+  if (options.enable && options.disable) {
+    return handleError(
+      "Cannot both enable and disable a feature.",
+      "Feature Access",
+    );
+  }
+
+  if (!options.enable && !options.disable) {
+    return handleError(
+      "Must specify either --enable or --disable.",
+      "Feature Access",
+    );
+  }
+
+  // Validate at least one target is specified
+  if (
+    !options.companies?.length &&
+    !options.segments?.length &&
+    !options.users?.length
+  ) {
+    return handleError(
+      "Must specify at least one target using --companies, --segments, or --users.",
+      "Feature Access",
+    );
+  }
+
+  // If feature key is not provided, let user select one
+  if (!featureKey) {
+    try {
+      spinner = ora(
+        `Loading features for app ${chalk.cyan(app.name)}${baseUrlSuffix(
+          baseUrl,
+        )}...`,
+      ).start();
+
+      const featuresResponse = await listFeatures(appId, {
+        envId: production.id,
+      });
+
+      if (featuresResponse.data.length === 0) {
+        return handleError("No features found for this app.", "Feature Access");
+      }
+
+      spinner.succeed(
+        `Loaded features for app ${chalk.cyan(app.name)}${baseUrlSuffix(baseUrl)}.`,
+      );
+
+      featureKey = await select({
+        message: "Select a feature to manage access:",
+        choices: featuresResponse.data.map((feature) => ({
+          name: `${feature.name} (${feature.key})`,
+          value: feature.key,
+        })),
+      });
+    } catch (error) {
+      spinner?.fail("Loading features failed.");
+      return handleError(error, "Feature Access");
+    }
+  }
+
+  // Determine if enabling or disabling
+  const isEnabled = options.enable;
+
+  const targets = [
+    ...(options.users?.map((id) => chalk.cyan(`user ID ${id}`)) ?? []),
+    ...(options.companies?.map((id) => chalk.cyan(`company ID ${id}`)) ?? []),
+    ...(options.segments?.map((id) => chalk.cyan(`segment ID ${id}`)) ?? []),
+  ];
+
+  try {
+    spinner = ora(
+      `${isEnabled ? "Enabling" : "Disabling"} feature ${chalk.cyan(featureKey)} for ${lf.format(targets)}...`,
+    ).start();
+
+    await updateFeatureAccess(appId, {
+      envId: production.id,
+      featureKey,
+      isEnabled,
+      companyIds: options.companies,
+      segmentIds: options.segments,
+      userIds: options.users,
+    });
+
+    spinner.succeed(
+      `${isEnabled ? "Enabled" : "Disabled"} feature ${chalk.cyan(featureKey)} for specified ${lf.format(targets)}.`,
+    );
+  } catch (error) {
+    spinner?.fail(`Feature access update failed.`);
+    void handleError(error, "Feature Access");
+  }
+};
+
 export function registerFeatureCommands(cli: Command) {
   const featuresCommand = new Command("features").description(
     "Manage features.",
@@ -199,6 +327,26 @@ export function registerFeatureCommands(cli: Command) {
     .addOption(typesOutOption)
     .addOption(typesFormatOption)
     .action(generateTypesAction);
+
+  // Add feature access command
+  featuresCommand
+    .command("access")
+    .description(
+      "Grant or revoke feature access for companies, segments, and users.",
+    )
+    .addOption(appIdOption)
+    .addArgument(
+      new Argument(
+        "[featureKey]",
+        "Feature key. If not provided, you'll be prompted to select one",
+      ),
+    )
+    .addOption(enableFeatureOption)
+    .addOption(disableFeatureOption)
+    .addOption(companyIdsOption)
+    .addOption(segmentIdsOption)
+    .addOption(userIdsOption)
+    .action(featureAccessAction);
 
   // Update the config with the cli override values
   featuresCommand.hook("preAction", (_, command) => {
